Parameterize walk speed & bike speed
